requirements.txt
python-stdnum>=1.5
